+++ src/projectgoldstarsx/Education.java
@@ -1,6 +1,10 @@
 package projectgoldstarsx;
 import java.awt.GridLayout;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 import javax.swing.JLabel;
+import javax.swing.JMenu;
+import javax.swing.JMenuBar;
 public class Education
 {
     public static ProgramWindow educationFrame;
@@ -15,7 +19,7 @@ public class Education
         educationFrame = new ProgramWindow("Education");
         educationFrame.setSize(900 * ProjectGoldStarsX.multiplier, 475 * ProjectGoldStarsX.multiplier);
         educationFrame.setLayout(new GridLayout(3, 3));
-        educationFrame.setStandardMenuBar();
+        educationFrame.setJMenuBar(menuBar());
         educationFrame.add(Components.headerLabel("Math:"));
         educationFrame.add(Components.actionLabel("Practice Addition", new ListenersEducationMouse.PracticeAdditionListener()));
         educationFrame.add(Components.actionLabel("Practice Subtraction", new ListenersEducationMouse.PracticeSubtractionListener()));
@@ -26,5 +30,32 @@ public class Education
         educationFrame.add(Components.actionLabel("3-Sentence Story", new ListenersEducationMouse.Reading3SentencesListener()));
         educationFrame.add(Components.actionLabel("4-Sentence Story", new ListenersEducationMouse.Reading4SentencesListener()));
         educationFrame.makeVisible();
+    }
+    
+    private JMenuBar menuBar()
+    {
+        JMenuBar menuBar = new JMenuBar();
+        menuBar.setBackground(ProjectGoldStarsX.color1);
+        menuBar.add(educationFrame.getCloseButton());
+        menuBar.add(educationFrame.getMaximizeButton());
+        menuBar.add(educationFrame.getWindowMenu());
+        menuBar.add(moreMenu());
+        return menuBar;
+    }
+    
+    private JMenu moreMenu()
+    {
+        StandardMenu menu = new StandardMenu("More");
+        menu.add(Components.standardMenuItem("About Education", new AboutEducationListener()));
+        return menu.getMenu();
+    }
+    
+    public static class AboutEducationListener implements ActionListener
+    {
+        @Override
+        public void actionPerformed(ActionEvent e)
+        {
+            AboutPrograms.aboutEducation();
+        }
     }
 }